Rename `AsyncStorage` to `storageUtils`
The old name did not really fit anymore, because the storage functionality has moved and async storage only contained utility functions.

Also, ideally, we should probably extract the functionality into different services and get rid of this file. This has low priority though.
