+++ src/components/__tests__/FavoriteButton.spec.tsx
@@ -3,8 +3,8 @@ import { fireEvent, waitFor } from '@testing-library/react-native'
 import React from 'react'
 
 import { VOCABULARY_ITEM_TYPES } from '../../constants/data'
-import { isFavorite } from '../../services/AsyncStorage'
 import { StorageCache } from '../../services/Storage'
+import { isFavorite } from '../../services/storageUtils'
 import VocabularyItemBuilder from '../../testing/VocabularyItemBuilder'
 import { renderWithStorageCache } from '../../testing/render'
 import FavoriteButton from '../FavoriteButton'
