+++ src/routes/choice-exercises/__tests__/ArticleChoiceExerciseScreen.spec.tsx
@@ -4,9 +4,9 @@ import React from 'react'
 
 import { ExerciseKeys, SIMPLE_RESULTS } from '../../../constants/data'
 import { RoutesParams } from '../../../navigation/NavigationTypes'
-import { saveExerciseProgress } from '../../../services/AsyncStorage'
 import { StorageCache } from '../../../services/Storage'
 import { getLabels } from '../../../services/helpers'
+import { saveExerciseProgress } from '../../../services/storageUtils'
 import VocabularyItemBuilder from '../../../testing/VocabularyItemBuilder'
 import createNavigationMock from '../../../testing/createNavigationPropMock'
 import render, { renderWithStorageCache } from '../../../testing/render'
@@ -22,7 +22,7 @@ jest.mock('../../../services/helpers', () => ({
   shuffleArray: jest.fn(it => it),
 }))
 
-jest.mock('../../../services/AsyncStorage', () => ({
+jest.mock('../../../services/storageUtils', () => ({
   getExerciseProgress: jest.fn(() => ({})),
   saveExerciseProgress: jest.fn(),
   getDevMode: jest.fn(async () => false),
