+++ src/routes/choice-exercises/components/SingleChoiceExercise.tsx
@@ -20,9 +20,9 @@ import {
 } from '../../../constants/data'
 import { AlternativeWord, VocabularyItem } from '../../../constants/endpoints'
 import { RoutesParams, VocabularyItemResult } from '../../../navigation/NavigationTypes'
-import { saveExerciseProgress } from '../../../services/AsyncStorage'
 import { StorageCacheContext } from '../../../services/Storage'
 import { calculateScore, getLabels, moveToEnd, shuffleArray, willNextExerciseUnlock } from '../../../services/helpers'
+import { saveExerciseProgress } from '../../../services/storageUtils'
 import { SingleChoice } from './SingleChoice'
 
 const ButtonContainer = styled.View`
