+++ src/routes/home/components/DisciplineCard.tsx
@@ -11,9 +11,9 @@ import { BUTTONS_THEME, NextExerciseData } from '../../../constants/data'
 import { Discipline, ForbiddenError, NetworkError } from '../../../constants/endpoints'
 import { isTypeLoadProtected } from '../../../hooks/helpers'
 import { RequestParams, useLoadDiscipline } from '../../../hooks/useLoadDiscipline'
-import { removeCustomDiscipline, removeSelectedProfession } from '../../../services/AsyncStorage'
 import { StorageCacheContext } from '../../../services/Storage'
 import { getLabels } from '../../../services/helpers'
+import { removeCustomDiscipline, removeSelectedProfession } from '../../../services/storageUtils'
 import Card from './Card'
 import CustomDisciplineDetails from './CustomDisciplineDetails'
 import ProfessionDetails from './ProfessionDetails'
