+++ src/routes/manage-selections/__tests__/ManageSelectionsScreen.spec.tsx
@@ -3,9 +3,9 @@ import { mocked } from 'jest-mock'
 import React from 'react'
 
 import { useLoadDiscipline } from '../../../hooks/useLoadDiscipline'
-import { pushSelectedProfession } from '../../../services/AsyncStorage'
 import { StorageCache } from '../../../services/Storage'
 import { getLabels } from '../../../services/helpers'
+import { pushSelectedProfession } from '../../../services/storageUtils'
 import createNavigationMock from '../../../testing/createNavigationPropMock'
 import { getReturnOf } from '../../../testing/helper'
 import { mockCustomDiscipline } from '../../../testing/mockCustomDiscipline'
