+++ src/routes/process-user-vocabulary/__tests__/UserVocabularyProcessScreen.spec.tsx
@@ -6,9 +6,9 @@ import ReactNativeFS, { DocumentDirectoryPath } from 'react-native-fs'
 import { ARTICLES, VOCABULARY_ITEM_TYPES } from '../../../constants/data'
 import { VocabularyItem } from '../../../constants/endpoints'
 import { RoutesParams } from '../../../navigation/NavigationTypes'
-import { getUserVocabularyItems } from '../../../services/AsyncStorage'
 import { StorageCache } from '../../../services/Storage'
 import { getLabels } from '../../../services/helpers'
+import { getUserVocabularyItems } from '../../../services/storageUtils'
 import createNavigationMock from '../../../testing/createNavigationPropMock'
 import render, { renderWithStorageCache } from '../../../testing/render'
 import UserVocabularyProcessScreen from '../UserVocabularyProcessScreen'
