+++ src/routes/user-vocabulary-list/UserVocabularyListScreen.tsx
@@ -15,10 +15,10 @@ import { VocabularyItem } from '../../constants/endpoints'
 import useKeyboard from '../../hooks/useKeyboard'
 import useReadUserVocabulary from '../../hooks/useReadUserVocabulary'
 import { RoutesParams } from '../../navigation/NavigationTypes'
-import { deleteUserVocabularyItem } from '../../services/AsyncStorage'
 import { StorageCacheContext } from '../../services/Storage'
 import { getLabels, getSortedAndFilteredVocabularyItems } from '../../services/helpers'
 import { reportError } from '../../services/sentry'
+import { deleteUserVocabularyItem } from '../../services/storageUtils'
 import ListEmptyContent from './components/ListEmptyContent'
 import ListItem from './components/ListItem'
